Simplify and make more flexible the logging API.
Changed the logging function to take an object with a well known interface rather than passing in a "enum" with a specific datatype associated with that enum value. This also allows the logging function to be more flexible, as it isn't limited to either logging a string message or a set of key-value pairs, but can compose a log message out of both of those things simultaneously.

Moved the query string parsing for debug logging activation in to log.js to keep the code clean.